# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost
  description: Generated server url
paths:
...
  /api/latest/user-info/coach:
    post:
      tags:
      - user-info-controller
      operationId: setCoach
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetCoachRequestDto"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserInfoDto"
...
  /api/latest/coach-availability/recurring:
    get:
      tags:
      - coach-availability-controller
      operationId: getRecurringCoachAvailability
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReoccurringEventDto"
    post:
      tags:
      - coach-availability-controller
      operationId: setRecurringCoachAvailability
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/ReoccurringEventDto"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
  /api/latest/coach-availability/non-recurring:
    get:
      tags:
      - coach-availability-controller
      operationId: getNonRecurringCoachAvailability
      parameters:
      - name: filterDto
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/EventFilterDto"
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NonReoccurringEventDto"
    post:
      tags:
      - coach-availability-controller
      operationId: setNonRecurringCoachAvailability
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetNonrecurringRequestDto"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
...
  /api/latest/coach-availability/recurring/settings:
    get:
      tags:
      - coach-availability-controller
      operationId: getRecurringSetting
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReoccurringAvailabilityDto"
    patch:
      tags:
      - coach-availability-controller
      operationId: updateRecurringSetting
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReoccurringAvailabilityDto"
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
...
  /api/latest/coach-availability/event-types:
    get:
      tags:
      - coach-availability-controller
      operationId: getCalendlyEvents
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ErrorDto"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventType"
...
components:
  schemas:
    ErrorCodeFieldDto:
      type: object
      properties:
        name:
          type: string
        value:
          type: string
    ErrorDto:
      type: object
      properties:
        errorCode:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/ErrorCodeFieldDto"
        errorMessage:
          type: string
...
    UserInfoDto:
      type: object
      properties:
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        userRoles:
          type: array
          items:
            type: string
            enum:
            - RESPONDENT
            - USER
            - MANAGER
            - COACH
            - HR
            - ADMIN
          uniqueItems: true
        timeZone:
          type: string
        strengths:
          type: array
          items:
            type: string
        values:
          type: array
          items:
            type: string
        areaOfDevelopment:
          type: array
          items:
            type: string
        coach:
          type: string
        locale:
          type: string
        allowedCoachRates:
          type: array
          items:
            type: string
          uniqueItems: true
        companyId:
          type: integer
          format: int64
...
    SetCoachRequestDto:
      type: object
      properties:
        coach:
          type: string
...
    ReoccurringEventDto:
      type: object
      properties:
        from:
          $ref: "#/components/schemas/ReoccurringEventTimeDto"
        to:
          $ref: "#/components/schemas/ReoccurringEventTimeDto"
    ReoccurringEventTimeDto:
      type: object
      properties:
        day:
          type: string
          enum:
          - MONDAY
          - TUESDAY
          - WEDNESDAY
          - THURSDAY
          - FRIDAY
          - SATURDAY
          - SUNDAY
        time:
          type: string
    EventFilterDto:
      type: object
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
      required:
      - from
      - to
    NonReoccurringEventDto:
      type: object
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
    SetNonrecurringRequestDto:
      type: object
      properties:
        timeFrame:
          $ref: "#/components/schemas/EventFilterDto"
        events:
          type: array
          items:
            $ref: "#/components/schemas/NonReoccurringEventDto"
      required:
      - events
      - timeFrame
...
    ReoccurringAvailabilityDto:
      type: object
      properties:
        uuid:
          type: string
        type:
          type: string
          enum:
          - GOOGLE
          - CALENDLY
          - CUSTOM
        active:
          type: boolean
...
    EventType:
      type: object
      properties:
        name:
          type: string
        uuid:
          type: string
        active:
          type: boolean
...